+++ VKTS/include/vkts/core/container/List.hpp
@@ -65,7 +65,7 @@ private:
     ListElement<V>* front;
     ListElement<V>* back;
 
-    size_t listSize;
+    uint32_t listSize;
 
 public:
 
@@ -219,7 +219,7 @@ public:
         listSize++;
     }
 
-    VkBool32 insert(const size_t index, const V& value)
+    VkBool32 insert(const uint32_t index, const V& value)
     {
         auto walker = front;
         auto counter = 0;
@@ -411,7 +411,7 @@ public:
         return VK_FALSE;
     }
 
-    V& operator[](const size_t index)
+    V& operator[](const uint32_t index)
     {
     	if (index > listSize)
     	{
@@ -438,7 +438,7 @@ public:
         return walker->element;
     }
 
-    const V& operator[](const size_t index) const
+    const V& operator[](const uint32_t index) const
     {
     	if (index >= listSize)
     	{
@@ -458,7 +458,7 @@ public:
         return walker->element;
     }
 
-    size_t size() const
+    uint32_t size() const
     {
         return listSize;
     }
