+++ VKTS/include/vkts/core/container/Map.hpp
@@ -48,7 +48,7 @@ public:
     {
     }
 
-    Map(const size_t& allDataCount) :
+    Map(const uint32_t& allDataCount) :
         allKeys(allDataCount), allValues(allDataCount)
     {
     }
@@ -89,12 +89,12 @@ public:
         allValues.clear();
     }
 
-    size_t find(const K& key) const
+    uint32_t find(const K& key) const
     {
-        size_t start = 0;
-        size_t end = allKeys.size();
+        uint32_t start = 0;
+        uint32_t end = allKeys.size();
 
-        size_t result = 0;
+        uint32_t result = 0;
 
         while (start < end)
         {
@@ -124,7 +124,7 @@ public:
 
     VkBool32 set(const K& key, const V& value)
     {
-        for (size_t i = 0; i < allKeys.size(); i++)
+        for (uint32_t i = 0; i < allKeys.size(); i++)
         {
             if (key == allKeys[i])
             {
@@ -149,7 +149,7 @@ public:
 
     VkBool32 remove(const K& key)
     {
-        size_t index = find(key);
+        uint32_t index = find(key);
 
         if (index != allKeys.size())
         {
@@ -159,7 +159,7 @@ public:
         return VK_FALSE;
     }
 
-    VkBool32 removeAt(const size_t index)
+    VkBool32 removeAt(const uint32_t index)
     {
         if (index >= allKeys.size())
         {
@@ -179,7 +179,7 @@ public:
 
     V& operator[](const K& key)
     {
-        size_t index = find(key);
+        uint32_t index = find(key);
 
         if (index != allKeys.size())
         {
@@ -197,22 +197,22 @@ public:
         return allValues[find(key)];
     }
 
-    K& keyAt(const size_t index)
+    K& keyAt(const uint32_t index)
     {
         return allKeys[index];
     }
 
-    const K& keyAt(const size_t index) const
+    const K& keyAt(const uint32_t index) const
     {
         return allKeys[index];
     }
 
-    V& valueAt(const size_t index)
+    V& valueAt(const uint32_t index)
     {
         return allValues[index];
     }
 
-    const V& valueAt(const size_t index) const
+    const V& valueAt(const uint32_t index) const
     {
         return allValues[index];
     }
@@ -227,7 +227,7 @@ public:
         return allValues;
     }
 
-    size_t size() const
+    uint32_t size() const
     {
         return allKeys.size();
     }
