+++ VKTS/include/vkts/core/container/SmartPointerVector.hpp
@@ -40,12 +40,12 @@ protected:
 
     V* allData;
 
-    size_t topElement;
-    size_t allDataCount;
+    uint32_t topElement;
+    uint32_t allDataCount;
 
-    size_t getAllocSize()
+    uint32_t getAllocSize()
     {
-    	size_t tempAllDataCount = allDataCount * 2;
+    	uint32_t tempAllDataCount = allDataCount * 2;
 
     	return tempAllDataCount > 0 ? tempAllDataCount : 1;
     }
@@ -58,7 +58,7 @@ public:
 
     }
 
-    SmartPointerVector(const size_t& allDataCount) :
+    SmartPointerVector(const uint32_t& allDataCount) :
         allData(nullptr), topElement(0), allDataCount(0)
     {
         allData = new V[allDataCount];
@@ -68,7 +68,7 @@ public:
         	throw std::bad_alloc();
         }
 
-        for (size_t i = 0; i < allDataCount; i++)
+        for (uint32_t i = 0; i < allDataCount; i++)
         {
             allData[i] = nullptr;
         }
@@ -87,7 +87,7 @@ public:
         	throw std::bad_alloc();
         }
 
-        for (size_t i = 0; i < other.allDataCount; i++)
+        for (uint32_t i = 0; i < other.allDataCount; i++)
         {
             allData[i] = nullptr;
             allData[i] = other.allData[i];
@@ -128,7 +128,7 @@ public:
         	throw std::bad_alloc();
         }
 
-        for (size_t i = 0; i < other.allDataCount; i++)
+        for (uint32_t i = 0; i < other.allDataCount; i++)
         {
             allData[i] = nullptr;
             allData[i] = other.allData[i];
@@ -188,7 +188,7 @@ public:
 
     void clear()
     {
-        for (size_t i = 0; i < topElement; i++)
+        for (uint32_t i = 0; i < topElement; i++)
         {
             allData[i].reset();
         }
@@ -207,7 +207,7 @@ public:
             	throw std::bad_alloc();
             }
 
-            for (size_t i = 0; i < getAllocSize(); i++)
+            for (uint32_t i = 0; i < getAllocSize(); i++)
             {
                 newAllData[i] = nullptr;
 
@@ -229,7 +229,7 @@ public:
         topElement++;
     }
 
-    VkBool32 insert(const size_t index, const V& value)
+    VkBool32 insert(const uint32_t index, const V& value)
     {
         if (index > topElement)
         {
@@ -252,7 +252,7 @@ public:
             	throw std::bad_alloc();
             }
 
-            for (size_t i = 0; i < getAllocSize(); i++)
+            for (uint32_t i = 0; i < getAllocSize(); i++)
             {
                 newAllData[i] = nullptr;
 
@@ -270,7 +270,7 @@ public:
             allDataCount = getAllocSize();
         }
 
-        for (size_t copyIndex = topElement; copyIndex >= index + 1; copyIndex--)
+        for (uint32_t copyIndex = topElement; copyIndex >= index + 1; copyIndex--)
         {
             allData[copyIndex] = allData[copyIndex - 1];
         }
@@ -283,7 +283,7 @@ public:
 
     VkBool32 remove(const V& value)
     {
-        for (size_t i = 0; i < topElement; i++)
+        for (uint32_t i = 0; i < topElement; i++)
         {
             if (allData[i] == value)
             {
@@ -298,7 +298,7 @@ public:
                     return VK_TRUE;
                 }
 
-                for (size_t copyIndex = i; copyIndex < topElement - 1;
+                for (uint32_t copyIndex = i; copyIndex < topElement - 1;
                         copyIndex++)
                 {
                     allData[copyIndex] = allData[copyIndex + 1];
@@ -314,7 +314,7 @@ public:
         return VK_FALSE;
     }
 
-    VkBool32 removeAt(const size_t index)
+    VkBool32 removeAt(const uint32_t index)
     {
         if (index >= topElement)
         {
@@ -326,7 +326,7 @@ public:
 
     VkBool32 contains(const V& value) const
     {
-        for (size_t i = 0; i < topElement; i++)
+        for (uint32_t i = 0; i < topElement; i++)
         {
             if (value == allData[i])
             {
@@ -337,7 +337,7 @@ public:
         return VK_FALSE;
     }
 
-    V& operator[](const size_t index)
+    V& operator[](const uint32_t index)
     {
     	if (index > topElement)
     	{
@@ -354,7 +354,7 @@ public:
         return allData[index];
     }
 
-    const V& operator[](const size_t index) const
+    const V& operator[](const uint32_t index) const
     {
     	if (index >= topElement)
     	{
@@ -369,14 +369,14 @@ public:
         return allData;
     }
 
-    size_t size() const
+    uint32_t size() const
     {
         return topElement;
     }
 
-    size_t index(const V& value) const
+    uint32_t index(const V& value) const
     {
-        for (size_t i = 0; i < topElement; i++)
+        for (uint32_t i = 0; i < topElement; i++)
         {
             if (value == allData[i])
             {
