+++ VKTS/include/vkts/scenegraph/scene/IScene.hpp
@@ -60,7 +60,7 @@ public:
 
     virtual IObjectSP findObject(const std::string& name) const = 0;
 
-    virtual size_t getNumberObjects() const = 0;
+    virtual uint32_t getNumberObjects() const = 0;
 
     virtual const SmartPointerVector<IObjectSP>& getObjects() const = 0;
 
@@ -71,7 +71,7 @@ public:
 
     virtual ICameraSP findCamera(const std::string& name) const = 0;
 
-    virtual size_t getNumberCameras() const = 0;
+    virtual uint32_t getNumberCameras() const = 0;
 
     virtual const SmartPointerVector<ICameraSP>& getCameras() const = 0;
 
@@ -82,7 +82,7 @@ public:
 
     virtual ILightSP findLight(const std::string& name) const = 0;
 
-    virtual size_t getNumberLights() const = 0;
+    virtual uint32_t getNumberLights() const = 0;
 
     virtual const SmartPointerVector<ILightSP>& getLights() const = 0;
 
@@ -105,13 +105,13 @@ public:
 
     //
 
-    virtual void updateParameterRecursive(const Parameter* parameter, const uint32_t objectOffset = 0, const uint32_t objectStep = 1, const size_t objectLimit = SIZE_MAX) = 0;
+    virtual void updateParameterRecursive(const Parameter* parameter, const uint32_t objectOffset = 0, const uint32_t objectStep = 1, const uint32_t objectLimit = UINT32_MAX) = 0;
 
-    virtual void updateDescriptorSetsRecursive(const uint32_t allWriteDescriptorSetsCount, VkWriteDescriptorSet* allWriteDescriptorSets, const uint32_t currentBuffer, const uint32_t objectOffset = 0, const uint32_t objectStep = 1, const size_t objectLimit = SIZE_MAX) = 0;
+    virtual void updateDescriptorSetsRecursive(const uint32_t allWriteDescriptorSetsCount, VkWriteDescriptorSet* allWriteDescriptorSets, const uint32_t currentBuffer, const uint32_t objectOffset = 0, const uint32_t objectStep = 1, const uint32_t objectLimit = UINT32_MAX) = 0;
 
-    virtual void updateTransformRecursive(const double deltaTime, const uint64_t deltaTicks, const double tickTime, const uint32_t currentBuffer = 0, const uint32_t objectOffset = 0, const uint32_t objectStep = 1, const size_t objectLimit = SIZE_MAX) = 0;
+    virtual void updateTransformRecursive(const double deltaTime, const uint64_t deltaTicks, const double tickTime, const uint32_t currentBuffer = 0, const uint32_t objectOffset = 0, const uint32_t objectStep = 1, const uint32_t objectLimit = UINT32_MAX) = 0;
 
-    virtual void drawRecursive(const ICommandBuffersSP& cmdBuffer, const SmartPointerVector<IGraphicsPipelineSP>& allGraphicsPipelines, const uint32_t currentBuffer, const std::map<uint32_t, VkTsDynamicOffset>& dynamicOffsetMappings, const OverwriteDraw* renderOverwrite = nullptr, const uint32_t objectOffset = 0, const uint32_t objectStep = 1, const size_t objectLimit = SIZE_MAX) = 0;
+    virtual void drawRecursive(const ICommandBuffersSP& cmdBuffer, const SmartPointerVector<IGraphicsPipelineSP>& allGraphicsPipelines, const uint32_t currentBuffer, const std::map<uint32_t, VkTsDynamicOffset>& dynamicOffsetMappings, const OverwriteDraw* renderOverwrite = nullptr, const uint32_t objectOffset = 0, const uint32_t objectStep = 1, const uint32_t objectLimit = UINT32_MAX) = 0;
 
 };
